+++ harbour/contrib/rddads/adsfunc.c
@@ -84,17 +84,17 @@ char * hb_adsOemToAnsi( const char * pszSrc, HB_SIZE nLen )
 {
    if( hb_ads_bOEM )
    {
-      int nWideLen = MultiByteToWideChar( CP_OEMCP, MB_PRECOMPOSED, pszSrc, nLen, NULL, 0 );
+      int nWideLen = MultiByteToWideChar( CP_OEMCP, MB_PRECOMPOSED, pszSrc, ( int ) nLen, NULL, 0 );
       LPWSTR pszWide = ( LPWSTR ) hb_xgrab( ( nWideLen + 1 ) * sizeof( wchar_t ) );
 
       char * pszDst;
 
-      MultiByteToWideChar( CP_OEMCP, MB_PRECOMPOSED, pszSrc, nLen, pszWide, nWideLen );
+      MultiByteToWideChar( CP_OEMCP, MB_PRECOMPOSED, pszSrc, ( int ) nLen, pszWide, nWideLen );
 
       nLen = WideCharToMultiByte( CP_ACP, 0, pszWide, nWideLen, NULL, 0, NULL, NULL );
       pszDst = ( char * ) hb_xgrab( nLen + 1 );
 
-      WideCharToMultiByte( CP_ACP, 0, pszWide, nWideLen, pszDst, nLen, NULL, NULL );
+      WideCharToMultiByte( CP_ACP, 0, pszWide, nWideLen, pszDst, ( int ) nLen, NULL, NULL );
 
       hb_xfree( pszWide );
 
@@ -108,17 +108,17 @@ char * hb_adsAnsiToOem( const char * pszSrc, HB_SIZE nLen )
 {
    if( hb_ads_bOEM )
    {
-      int nWideLen = MultiByteToWideChar( CP_ACP, MB_PRECOMPOSED, pszSrc, nLen, NULL, 0 );
+      int nWideLen = MultiByteToWideChar( CP_ACP, MB_PRECOMPOSED, pszSrc, ( int ) nLen, NULL, 0 );
       LPWSTR pszWide = ( LPWSTR ) hb_xgrab( ( nWideLen + 1 ) * sizeof( wchar_t ) );
 
       char * pszDst;
 
-      MultiByteToWideChar( CP_ACP, MB_PRECOMPOSED, pszSrc, nLen, pszWide, nWideLen );
+      MultiByteToWideChar( CP_ACP, MB_PRECOMPOSED, pszSrc, ( int ) nLen, pszWide, nWideLen );
 
       nLen = WideCharToMultiByte( CP_OEMCP, 0, pszWide, nWideLen, NULL, 0, NULL, NULL );
       pszDst = ( char * ) hb_xgrab( nLen + 1 );
 
-      WideCharToMultiByte( CP_OEMCP, 0, pszWide, nWideLen, pszDst, nLen, NULL, NULL );
+      WideCharToMultiByte( CP_OEMCP, 0, pszWide, nWideLen, pszDst, ( int ) nLen, NULL, NULL );
 
       hb_xfree( pszWide );
 
@@ -1594,7 +1594,7 @@ HB_FUNC( ADSVERSION )
          ucVersion[ 0 ] = '\0';
    }
 
-   iPos = strlen( ucVersion ) - 1;
+   iPos = ( int ) strlen( ucVersion ) - 1;
    while( iPos >= 0 && ucVersion[ iPos ] == ' ' )  /* remove trailing spaces */
       ucVersion[ iPos-- ] = '\0';
